backend/requirements.txt
fastapi
uvicorn
requests
google-generativeai
pydantic
ffmpeg-python
python-multipart
python-dotenv
